+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fix
+
+* Change order of event filtering mechanisms (#2001)
+
 ## 6.0.0-beta.2
 
 * Fix: Android profiling initializes on first profile start (#2009)